fix(core): PrimePower p/r 从 uint32_t 升级为 uint64_t，修复大素数截断
PrimePower::p 和 PrimePower::r 从 uint32_t 改为 uint64_t，解决两个关键
缩放问题：

1. cofactorizer::add_large_primes() 中 static_cast<uint32_t>(factor) 对
   >UINT32_MAX 的大素数静默截断为垃圾值，导致所有含大素数的关系数据损坏
2. FactorBaseParams::large_prime_bound 从 uint32_t 改为 uint64_t，
   消除从 GNFSParams(uint64_t) 复制时的截断

同时更新：PrimePowerHash 改用 FNV-1a 混合两个 uint64_t 字段，
FactorBase::Stats::large_prime_bound 匹配升级。

新增 3 组测试（17 个断言）覆盖：PrimePower uint64 存储、FactorBaseParams
uint64 lpb、关系序列化/反序列化大素数保真。

+++ include/gnfs/cofactor/cofactorizer.hpp
@@ -245,20 +245,16 @@ private:
 
         switch (cls.type) {
             case CofactorClass::Prime:
-                list.push_back(PrimePower{
-                    static_cast<uint32_t>(cls.factor1), 1});
+                list.push_back(PrimePower{cls.factor1, static_cast<uint8_t>(1)});
                 break;
 
             case CofactorClass::PrimePower:
-                list.push_back(PrimePower{
-                    static_cast<uint32_t>(cls.factor1), cls.power});
+                list.push_back(PrimePower{cls.factor1, cls.power});
                 break;
 
             case CofactorClass::Semiprime:
-                list.push_back(PrimePower{
-                    static_cast<uint32_t>(cls.factor1), 1});
-                list.push_back(PrimePower{
-                    static_cast<uint32_t>(cls.factor2), 1});
+                list.push_back(PrimePower{cls.factor1, static_cast<uint8_t>(1)});
+                list.push_back(PrimePower{cls.factor2, static_cast<uint8_t>(1)});
                 break;
 
             default:
